Rename the files as well
