+++ src/qt/qt_openglrenderer.hpp
